// src/types/payload.types.ts
export interface Blog {
  id: number
  slug: string
  title: string
  content: {
    root: {
      type: string
      children: {
        type: string
        version: number
        children?: {
          text: string
          children?: {
            [k: string]: string
          }[]
          fields?: {
            linkType: string
            newTab: string
            url: string
            [k: string]: string
          }
          [k: string]: string | any
        }[]
        [k: string]: any
      }[]
      value: {
        url: string
        [k: string]: any
      }
      direction: ('ltr' | 'rtl') | null
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | ''
      indent: number
      version: number
    }
    [k: string]: unknown
  }
  excerpt?: string | null
  isPublished?: boolean | null
  updatedAt: string
  createdAt: string
}